+++ monitor/services/src/main/java/org/datacleaner/monitor/configuration/ComponentCacheMapImpl.java
@@ -25,7 +25,6 @@ import java.util.List;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
-import org.datacleaner.configuration.RemoteComponentsConfiguration;
 import org.datacleaner.monitor.server.components.ComponentHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
