+++ monitor/services/src/main/java/org/datacleaner/monitor/configuration/RemoteComponentsConfiguration.java
@@ -1,0 +1,55 @@
+/**
+ * DataCleaner (community edition)
+ * Copyright (C) 2014 Neopost - Customer Information Management
+ *
+ * This copyrighted material is made available to anyone wishing to use, modify,
+ * copy, or redistribute it subject to the terms and conditions of the GNU
+ * Lesser General Public License, as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public License
+ * for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this distribution; if not, write to:
+ * Free Software Foundation, Inc.
+ * 51 Franklin Street, Fifth Floor
+ * Boston, MA  02110-1301  USA
+ */
+package org.datacleaner.monitor.configuration;
+
+import org.datacleaner.api.Component;
+import org.datacleaner.descriptors.ComponentDescriptor;
+
+/**
+ * Class RemoteComponentsConfiguration
+ *
+ */
+public interface RemoteComponentsConfiguration {
+
+
+    /**
+     * Is the component allowed?
+     *
+     * @param componentDescriptor
+     * @return
+     */
+    public boolean isAllowed(ComponentDescriptor componentDescriptor);
+
+    /**
+     * Is the component allowed?
+     *
+     * @param componentDisplayName
+     * @return
+     */
+    public boolean isAllowed(String componentDisplayName);
+
+    /**
+     * Method put or replaces the default values for specific controller.
+     *
+     * @param componentDescriptor
+     * @param component
+     */
+    public void setDefaultValues(ComponentDescriptor componentDescriptor, Component component);
+}
